requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-allauth-2fa==0.11.1
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-jalali-date==1.1.0
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-rosetta==0.10.0
django-view-breadcrumbs==2.5.1
factory-boy==3.3.0
Faker==23.2.1
idna==3.6
isodate==0.6.1
jdatetime==4.1.1
lxml==5.1.0
oauthlib==3.2.2
phonenumberslite==8.13.30
pillow==10.2.0
platformdirs==4.2.0
polib==1.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.0
zeep==4.2.1
